Updates debugging and fixes a connection issue
To use the debugging either you can set LOGGING.logging to True for
console output or you can also set a file like object as well.

There was an issue with being able to connect more then one pipe at the
same time. This has now been fixed. It had to do with how I was
handling the errors that are gotten from Windows as well as having one
of the error numbers wrong.